dashboard/requirements_dashboard.txt
alembic==0.9.1
aniso8601==1.2.0
appdirs==1.4.3
blinker==1.4
ciso8601==1.0.3
click==6.7
contextlib2==0.5.5
Flask==0.12.1
Flask-Cors==3.0.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.5
Flask-RQ2==17.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SSE==0.2.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.7.1
hashids==1.2.0
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
msgpack-python==0.4.8
olefile==0.44
packaging==16.8
passlib==1.7.1
Pillow==4.1.1
psycopg2==2.7.1
PyJWT==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
raven==6.0.0
requests==2.13.0
rq==0.7.1
rq-scheduler==0.7.0
six==1.10.0
SQLAlchemy==1.1.9
sqlalchemy-postgres-copy==0.5.0
unicodecsv==0.14.1
utm==0.4.2
Werkzeug==0.12.1
WTForms==2.1
